requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
Django==6.0.3
django-allauth==65.15.0
idna==3.11
pillow==12.1.1
pycparser==3.0
PyJWT==2.12.1
requests==2.33.0
sqlparse==0.5.5
urllib3==2.6.3
channels>=4.0.0
daphne>=4.0.0
whitenoise>=6.0.0
dj-database-url>=2.0.0
gunicorn>=21.2.0
uvicorn[standard]>=0.23.0
channels_redis>=4.1.0
psycopg2-binary>=2.9.0
